[Remove] 불필요 클래스 삭제 및 클래스 이름변경
