--- src/main/java/domain/Number.java
@@ -1,13 +1,0 @@
-package domain;
-
-public class Number {
-    private final int number;
-
-    public Number(int number) {
-        this.number = number;
-    }
-
-    public int value() {
-        return number;
-    }
-}
